Fix bootstrap servers by installing apparmor package
Probably because of a new version of docker-ce package
apparmor is not installed by default anymore and
containers can't start because of apparmor_parser is missing.

This patch is adding apparmor package to list of packages to
be installed.

Closes-Bug: #2004583
Depends-on: https://review.opendev.org/c/openstack/kolla-ansible/+/872558
Change-Id: I4855a9ea896f30575b924a45f020c23a2e780f4b
(cherry picked from commit fcb91f1e56faf4365b8ec856d945887526b9cd4a)