docker: Set ulimit nofile on EL9
Some latest RPM-based distributions, such as RHEL 9 and CentOS Stream 9, ship
a recent version of systemd that sets the default open file handle limit is
set to 1073741816.
That effects in various issues (like rabbitmq not working), including
ceph/other services in containers not working properly.
kolla-ansible has set that as a default ulimit for Kolla container images [1],
this patch sets that as a Docker Engine default (for container images
that don't have this overridden).

[1]: https://review.opendev.org/c/openstack/kolla-ansible/+/839715

Closes-Bug: #2008761

Change-Id: I48a8f18287aa3b017ce04d4b076d88989f927a56