Fix: include missing docker log-opts in config
The default docker config options, which specify the ``log-opts`` under
``roles/docker/vars/main.yml``, were not included in the
``docker_config`` combinations when the docker role was extracted from
the baremetal role in the Zed release. This means that
``docker_log_max_file`` and ``docker_log_max_size`` are not written into
the docker daemon.json.

The var name needs to be changed, as it can't be recursively defined
within the ``docker_config`` definition under
``roles/docker/tasks/config.yml`.

Closes-Bug: #2040105
Change-Id: I05efda2a9699336136dce8cfa99ada77b9b2a17b