Move podman socket activation to podman role
The expectation was that the podman role would fully configure the
podman service, but we needed to run the podman_sdk role for socket
activation to be enabled.

It also meant that you needed to run the podman_sdk role after the
podman role (as it expected that podman was already installed) which is
not always desirable. For instance, in kayobe, we setup the virtualenv
before configuring the container engine.

Closes-Bug: #2054568
Change-Id: Ic9d3cf9ad94f0cb28f9f1073ff02842a38211a54