Add installation of dbus package
In kolla-ansible, we perform a precheck to see if the python library for
dbus exists, but we do not install it in the bootstrap. This patch
addresses this issue by adding the installation of the dbus python
library.

Change-Id: Idfd4ffd0f80a2f22098639f81be69c664198b6c8
Closes-Bug: #2021585
(cherry picked from commit 1fd93d145a498e1d68e707701e6b3879a6505a58)