Allow creating /etc/hosts without Ansible facts
When fact gathering for specific groups is disabled, the
etc_hosts role does not have enough information to determine
either the IP address of the API network, or the Ansible
hostname/nodename of all hosts [1].

Allow operators to statically override the API network address
of a host in ``/etc/hosts`` by setting ``etc_hosts_api_address``
as an inventory variable. If specified, this takes precedence
over the API interface address currently used.

Also, if the ansible nodename and hostname facts are
not available for a host, fall-back to determining FQDN and
hostname from Ansible special variables [2], which do not
require fact gathering.

[1] https://docs.openstack.org/kolla-ansible/latest/user/ansible-tuning.html#delegated-fact-gathering
[2] https://docs.ansible.com/ansible/latest/reference_appendices/special_variables.html#special-variables

Change-Id: I53ed8b2f553c1eea263770b30860d224375b26b3