Fix bootstrapping in kolla-ansible
The idea behind kolla-ansible is to provide the ability to
bootstrap a server in such a way that kolla can orchestrate
the servers and operate against them. Historically, this part
of the code was moved from the kolla-ansible repository to the
ansible-collection-kolla repository, which is supposed to handle
the installation of podman/docker SDK and other dependencies -
whether through system packages or through a virtual environment.

However, the problem is that our CI never showed an error because
those dependencies were installed even before the actual kolla-ansible
bootstrap took place, but in real world this was not working.

This patch fixes exactly that: the dependencies that must be installed
beforehand so that other packages can be installed via pip are now handled
by ansible-collection-kolla. This code was removed from the Zull CI playbook,
as can be seen in the "needed-by" section below.

Needed-By: https://review.opendev.org/c/openstack/kolla-ansible/+/946963
Closes-Bug: #2109509

Change-Id: I288a32d46d2c18594cebf8f2fc40e6bd08607659
(cherry picked from commit ecad94400eb7e78f8f44d7909e034ec5b0558dd1)