Move tasks from k-a role common to a-c-k
Move tasks that modified host configuration from kolla-ansible
role common to a-c-k as they need to be run only once on the
bootstrap of the host and are not strongly related to the common
services.
Some extra options like mode and changed_when were added to keep
the linter happy.

Needed-by: https://review.opendev.org/c/openstack/kolla-ansible/+/948526
Change-Id: Ifc1b01a1acafc4064ccfad15cf2a691998209904
Signed-off-by: Roman Krček <roman.krcek@tietoevry.com>